backend/pyproject.toml
[tool.poetry]
name = "infograph-backend"
version = "0.1.0"
description = "Backend for Infograph Assistant"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95"
uvicorn = {extras = ["standard"], version = "^0.22"}
python-jose = "^3.3"
python-dotenv = "^1.0"
httpx = "^0.24"
pydantic = "^2.2"
pytest = {version = "^7.3", optional = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
